core: correct type of len in silkworm_rmd160

+++ silkworm/core/crypto/rmd160.c
@@ -362,13 +362,13 @@ static inline uint32_t load32(const void* src) {
     return w;
 }
 
-void silkworm_rmd160(uint8_t out[20], const uint8_t* ptr, size_t len) {
+void silkworm_rmd160(uint8_t out[20], const uint8_t* ptr, uint32_t len) {
     uint32_t buf[160 / 32];
 
     rmd160_init(buf);
 
     uint32_t current[16];
-    for (size_t remaining = len; remaining >= 64; remaining -= 64) {
+    for (unsigned remaining = len; remaining >= 64; remaining -= 64) {
         for (unsigned i = 0; i < 16; ++i) {
             current[i] = load32(ptr);
             ptr += 4;
